+++ src/docs/entries/concepts/lambdas/mod.rs
