+++ src/docs/entries/mod.rs
@@ -29,22 +29,22 @@ pub fn concept_entries() -> Vec<&'static ConceptEntry> {
     vec![
         &concepts::arrays::ARRAYS,
         &concepts::constants::CONSTANTS,
-        &concepts::flow::CONTROL_FLOW,
-        &concepts::flow::FOR_LOOPS,
+        &concepts::flow_control::CONTROL_FLOW,
+        &concepts::flow_control::FOR_LOOPS,
         &concepts::functions::FUNCTIONS,
-        &concepts::general::COMMENTS,
+        &concepts::generals::COMMENTS,
         &concepts::imports::IMPORTS,
         &concepts::lambdas::LAMBDAS,
-        &concepts::null::NULL,
+        &concepts::nulls::NULL,
         &concepts::operators::OPERATORS,
         &concepts::types::TYPES,
         &concepts::variables::VARIABLES,
-        &concepts::entry::ENTRY_POINTS,
+        &concepts::entries::ENTRY_POINTS,
         &concepts::tooling::TOOLING,
-        &concepts::byte::BYTES,
+        &concepts::bytes::BYTES,
         &concepts::tuples::TUPLES,
         &concepts::errors::ERROR_TYPE,
-        &concepts::casting::CASTING,
+        &concepts::casts::CASTING,
     ]
 }
